[CWS] move activity dump path utilities to separate package
