+++ pkg/security/resolvers/process/resolver_windows.go
@@ -22,7 +22,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/security/config"
 	"github.com/DataDog/datadog-agent/pkg/security/metrics"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
-	"github.com/DataDog/datadog-agent/pkg/security/utils"
+	"github.com/DataDog/datadog-agent/pkg/security/utils/pathutils"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -135,8 +135,8 @@ func (p *Resolver) AddNewEntry(pid uint32, ppid uint32, file string, envs []stri
 	e := p.processCacheEntryPool.Get()
 	e.PIDContext.Pid = pid
 	e.PPid = ppid
-	e.Process.CmdLine = utils.NormalizePath(commandLine)
-	e.Process.FileEvent.PathnameStr = utils.NormalizePath(file)
+	e.Process.CmdLine = pathutils.NormalizePath(commandLine)
+	e.Process.FileEvent.PathnameStr = pathutils.NormalizePath(file)
 	e.Process.FileEvent.BasenameStr = filepath.Base(e.Process.FileEvent.PathnameStr)
 	e.Process.EnvsEntry = &model.EnvsEntry{
 		Values: envs,
@@ -244,8 +244,8 @@ func (p *Resolver) Snapshot() {
 		e.PIDContext.Pid = Pid(pid)
 		e.PPid = Pid(proc.Ppid)
 
-		e.Process.CmdLine = utils.NormalizePath(strings.Join(proc.GetCmdline(), " "))
-		e.Process.FileEvent.PathnameStr = utils.NormalizePath(proc.Exe)
+		e.Process.CmdLine = pathutils.NormalizePath(strings.Join(proc.GetCmdline(), " "))
+		e.Process.FileEvent.PathnameStr = pathutils.NormalizePath(proc.Exe)
 		e.Process.FileEvent.BasenameStr = filepath.Base(e.Process.FileEvent.PathnameStr)
 		e.ExecTime = time.Unix(0, proc.Stats.CreateTime*int64(time.Millisecond))
 		entries = append(entries, e)
