+++ pkg/security/utils/pathutils/path_windows.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package utils
+package pathutils
 
 import (
 	"strings"
